scaffolding out the packager
