+++ packages/deployer-cf-workers/src/utils.ts
@@ -1,0 +1,3 @@
+import { _log } from '@fab/cli'
+
+export const log = _log(`@fab/deployer-cf-workers`)
